Allow using wrappers for objects. Might be tricky.
So if a key is an object responding to :call, then
it would be called to wrap the original object to be used
for the next coming schema. This way you could easily
adapt to whatever interface originally it is...

However it might not be so straightforward to use,
I could understand. That's the cost of trying to be
extremely flexible, unfortunately.